# Fix some type instabilities and add CI