--- a/Project.toml
+++ b/Project.toml
@@ -4,10 +4,13 @@ authors = ["Reinier Doelman <rdoelman@users.noreply.github.com>"]
 version = "0.1.0"
 
 [deps]
-Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [compat]
 julia = "1"
 
 [extras]
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
+
+
+[targets]
+test = ["Test"]