Hi!

I tried to fix some of the type preserving issues. Also, I worked on fixing a few Julia related performance bugs.
That improved performance quite a bit:

Current state:
```julia
julia> using ZernikePolynomials

julia> x = randn((2000,));

julia> y = x';

julia> using BenchmarkTools

julia> @btime Zernike(10, coord=:cartesian, index=:OSA).($x,$y);
  605.860 ms (16000026 allocations: 457.76 MiB)

# before it returned Float64
Zernike(10, coord=:cartesian, index=:OSA).(0.5f0, 0.5f0)
-6.911376921754665e-8
```

New state:
```julia
julia> using ZernikePolynomials
[ Info: Precompiling ZernikePolynomials [e462d300-c971-11e9-30f1-fb93b1f8f73a]

julia> x = randn((2000,));

julia> y = x';

julia> using BenchmarkTools

julia> @btime Zernike(10, coord=:cartesian, index=:OSA).($x,$y);
  165.865 ms (6 allocations: 30.52 MiB)

# now a Float32
julia> Zernike(10, coord=:cartesian, index=:OSA).(0.5f0, 0.5f0)
-6.911377f-8
```

Another change was to throw errors for invalid index combinations. In one `if-else` branch you returned an `Int` in the other one a `NaN` which is a `Float64` type. That's a type instability which I think is better solved by throwing an error.

Also improved `evaluateZernike` a little:
```julia
# OLD
julia> @btime ϕ = evaluateZernike(256, 2, 0.54, index=:OSA);
  4.940 ms (393231 allocations: 11.00 MiB)

julia> x = randn((100,)); 

julia> @btime ϕ = evaluateZernike(x, 2, 0.54, index=:OSA);

  816.550 μs (60015 allocations: 1.68 MiB)


# THIS PR
julia> @btime ϕ = evaluateZernike(256, 2, 0.54, index=:OSA);
 2.263 ms (15 allocations: 513.14 KiB)

julia> x = randn((100,));

julia> @btime ϕ = evaluateZernike(x, 2, 0.54, index=:OSA);
  328.311 μs (14 allocations: 78.69 KiB)

```



I would be happy to incorporate your feedback!

Best,

Felix