@roflmaostc 2022-06-03
Should we switch everything to 4 space?

There is another type stability issue because we return different functions at [this line](https://github.com/roflmaostc/ZernikePolynomials.jl/blob/a82953ca80e81d02b69b22154381fd0b5c77caf1/src/ZernikePolynomials.jl#L259).

I try to tackle that next week