@roflmaostc 2022-06-07
I think this type stability issue, we can leave it like that.
Otherwise, the code gets more complicated and less efficient in certain circumstances.